Slight movement refactoring.
This stops execution while the player is being pushed by a piston.  The
intent is that pistons happen first, although this will affect the cycle
lengths when the player is pushed.  However, it is generally the case
that the player is pushed into a pit and so the future cycles are
unimportant.