Remove player object from level.
The level screen object needs direct access to the player.  Rather than
exposing the player object as a public member and adding several
wrappers to the level object, the player should be an entirely separate
object that is managed by the level screen.

This also makes it easier to implement block puhsing since the
abstraction is more representative of the objects that will interact.